utrecht_test: Require person running test specify UTRECHT_TEST_DB
This way tests dont' confusingly fail with long errors simply because
somebody ran `boot test` without having previously configured a postgres
database.

This is the only sensible fence, requiring that they know there's a DB
setup somewhere.

Otherwise the defaults are mostly sensible, and are documented
explicitly at the top for easy grepping.

As a niceity, when the user does have UTRECHT_TEST_DB defined, a message
of:

  Testing vs 127.0.0.1:21314/databasename

Will be emitted at the top of the test run, to further help debugging.